add confirmation

+++ markov.py
@@ -17,9 +17,11 @@ bot = telebot.TeleBot(settings.TELEGRAM_TOKEN)
 def is_from_admin(message):
     username = message.from_user.username
     chat_id = str(message.chat.id)
-    username_admins = [
-        u.user.username for u in bot.get_chat_administrators(chat_id)
-    ]
+    username_admins = []
+    if message.chat.type != 'private':
+        username_admins = [
+            u.user.username for u in bot.get_chat_administrators(chat_id)
+        ]
     return (username in username_admins + settings.ADMIN_USERNAMES)
 
 
@@ -57,13 +59,28 @@ def generate_sentence(message, reply=False):
 @bot.message_handler(commands=['remove'])
 def remove_messages(message):
     if is_from_admin(message):
-        chat_id = str(message.chat.id)
-        db.delete(chat_id=chat_id)
-        get_model.cache_clear()
-        bot.reply_to(message, 'messages deleted')
-        logger.info(f'removing messages from {chat_id}')
-        return
+        if message.text == '/remove':
+            markup = telebot.types.ReplyKeyboardMarkup(
+                row_width=1, one_time_keyboard=True
+            )
+            markup.add('yes', 'no')
+            reply = bot.reply_to(
+                message, 'this operation will delete all data, are you sure?',
+                reply_markup=markup
+            )
+            bot.register_next_step_handler(reply, remove_messages)
 
+        elif message.text == 'yes':
+            chat_id = str(message.chat.id)
+            db.delete(chat_id=chat_id)
+            get_model.cache_clear()
+            bot.reply_to(message, 'messages deleted')
+            logger.info(f'removing messages from {chat_id}')
+
+        elif message.text == 'no':
+            bot.reply_to(message, 'aborting then')
+
+        return
     bot.reply_to(message, 'u r not an admin 🤔')
